Version 1.1 Release
What's New: Honestly nothing much, I still need to add the .py file, currently refactoring so it looks pretty. The jupyter notebook works now, had a large error in calculating the optimized weights but it is now fixed; max_sr_w is the array containing the optimal set of weights.